🐛 fix: resolve region stats panel showing No Data Available for all regions
- Extract ExplorePageContent into explore-page-client.tsx and simplify
  page.tsx to a thin server component; removes broken server-side prefetch
  that stored wrong-shaped data in the TanStack Query cache under the same
  key used by useRegionInfo
- Remove usePrefetchRegionInfo which poisoned the cache with raw API
  responses instead of processed RegionData, causing every hover→click
  to serve wrongly-shaped data (hasData = undefined → EmptyState shown)
- Fix YearOverYearBadge null guard: change === undefined to == null so
  null values from the RPC do not reach toFixed() causing a TypeError
- Fix DonutPieChart null guard: change !== undefined to != null for the
  same reason
